apps/worker/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
psycopg[binary]>=3.2.0
sqlalchemy>=2.0.0

# Job queue
bullmq>=2.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Cryptography
cryptography>=42.0.0

# Ads APIs
google-ads>=24.0.0

# Dev
pytest>=8.0.0
httpx>=0.27.0
